aider: Added support for passing tags as part of the system prompt.
# Aider chat conversation:

USER: I did `git reset --hard HEAD~1` to discard the last edits.

ASSISTANT: Ok.

USER: Pass in the tags as part of the system prompt

ASSISTANT: To pass in the tags as part of the system prompt, we need to modify the `systemPrompt` function in the `src/openai.ts` file. We will add a parameter for the tags and include them in the system prompt.

Here is the *edit block* for this change:

```typescript
src/openai.ts
<<<<<<< HEAD
const systemPrompt = (recentNotes) => {
    const notesString = recentNotes.slice(-5).map((note) => {
        return Object.entries(note.fields).map(([field, value]) => `${field}: ${value.value}`).join('\n');
    }).join('\n');
    return `You are an assistant assigned to create Anki cards.
You should make sure all cards are concise but have
enough context to understand seen out of context when reviewed
in the future. You don't need to be chatty, or have proper
grammar necessarily.

${notesString.length > 0 && "Here are some examples of cards we have:\n" + notesString}

Create cards based on the user's passed in prompt.`
};
=======
const systemPrompt = (recentNotes, tags) => {
    const notesString = recentNotes.slice(-5).map((note) => {
        return Object.entries(note.fields).map(([field, value]) => `${field}: ${value.value}`).join('\n');
    }).join('\n');
    return `You are an assistant assigned to create Anki cards.
You should make sure all cards are concise but have
enough context to understand seen out of context when reviewed
in the future. You don't need to be chatty, or have proper
grammar necessarily.

${notesString.length > 0 && "Here are some examples of cards we have:\n" + notesString}

Create cards based on the user's passed in prompt. The tags for these cards are: ${tags.join(', ')}`
};
>>>>>>> updated
```

This change will ensure that the tags are passed into the `systemPrompt` function and included in the system prompt. The `tags.join(', ')` is used to convert the tags array into a string with each tag separated by a comma and a space.