aider: Added filterOptions to support showing an "Add \"{tag}\"" dropdown in the Autocomplete component.
# Aider chat conversation:

USER: Use filterOption to support showing an "Add \"{tag}\"" dropdown

ASSISTANT: To support showing an "Add \"{tag}\"" dropdown, we need to modify the `Autocomplete` component in the `src/App.tsx` file. We will add the `filterOptions` prop to the `Autocomplete` component. This prop will be a function that takes the options, the state of the input, and returns the filtered options. If the input value doesn't match any of the options, it will return an option to add the input value.

Here is the *edit block* for this change:

```tsx
src/App.tsx
<<<<<<< HEAD
<Autocomplete
    id="tags"
    multiple
    autoHighlight
    value={currentTags}
    options={tags || []}
    onChange={(_, value) => { value && setCurrentTags(value) }}
    renderInput={(params) => <TextField label="Tags" {...params} />}
/>
=======
<Autocomplete
    id="tags"
    multiple
    autoHighlight
    value={currentTags}
    options={tags || []}
    onChange={(_, value) => { value && setCurrentTags(value) }}
    filterOptions={(options, params) => {
        const filtered = options.filter((option) =>
            option.toLowerCase().includes(params.inputValue.toLowerCase())
        );

        if (params.inputValue !== '') {
            filtered.push(`Add "${params.inputValue}"`);
        }

        return filtered;
    }}
    renderInput={(params) => <TextField label="Tags" {...params} />}
/>
>>>>>>> updated
```

This change ensures that the `Autocomplete` component includes an "Add \"{tag}\"" option in the dropdown if the input value doesn't match any of the existing options. The input value is added to the filtered options if it's not an empty string.