aider: Added an AppBar with a Toolbar and Typography component to the Root component.
# Aider chat conversation:

USER: OK, take the Root component, and add in an AppBar

ASSISTANT: To add an AppBar to the Root component, we will need to modify the `src/main.tsx` file. We will import the AppBar component from the Material-UI library and add it to the Root component.

Here is the edit block for the changes:

```tsx
src/main.tsx
<<<<<<< HEAD
import { Outlet } from 'react-router-dom';

function Root() {
    return <Outlet />;
}
=======
import { AppBar, Toolbar, Typography } from '@mui/material';
import { Outlet } from 'react-router-dom';

function Root() {
    return (
        <>
            <AppBar position="static">
                <Toolbar>
                    <Typography variant="h6" component="div">
                        My App
                    </Typography>
                </Toolbar>
            </AppBar>
            <Outlet />
        </>
    );
}
>>>>>>> updated
```

This edit block adds an AppBar with a Toolbar and a Typography component for the title of the app. The Outlet component is wrapped in a fragment along with the AppBar.