aider: Add a link to the home page on the "anki card creator" typography.
# Aider chat conversation:

USER: Add a link to home to the anki card creator typography thing

ASSISTANT: To add a link to the home page on the "anki card creator" typography, we will need to modify the `src/main.tsx` file.

Here are the steps:

1. Import the `Link` component from `react-router-dom`.
2. Wrap the `Typography` component that contains "anki card creator" with the `Link` component and set the `to` prop to "/" (the home page).

Here is the *edit block* for these changes:

```tsx
src/main.tsx
<<<<<<< HEAD
<Typography variant="h6" component="div">anki card creator</Typography>
=======
<Link to="/">
    <Typography variant="h6" component="div">anki card creator</Typography>
</Link>
>>>>>>> updated
```
This will make the "anki card creator" text clickable and redirect to the home page when clicked.