aider: Make "Anki Connect" a link to https://ankiweb.net/shared/info/2055492159.
# Aider chat conversation:

USER: anki connect should link to this page https://ankiweb.net/shared/info/2055492159

ASSISTANT: To make "Anki Connect" a link, we will use the `Link` component from the `@mui/material` library. We will wrap the text "Anki Connect" with the `Link` component and set the `href` attribute to "https://ankiweb.net/shared/info/2055492159".

Here is the *edit block* for the changes:

```tsx
src/Home.tsx
<<<<<<< HEAD
            <Typography variant="body1" sx={{ marginTop: 2 }}>
                2. Have Anki Connect plugin installed in your Anki instance
            </Typography>
=======
            <Typography variant="body1" sx={{ marginTop: 2 }}>
                2. Have <Link href="https://ankiweb.net/shared/info/2055492159" target="_blank" rel="noopener noreferrer">Anki Connect</Link> plugin installed in your Anki instance
            </Typography>
>>>>>>> updated
```

This will make "Anki Connect" a link that opens in a new tab when clicked.